requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.8
Django==5.2.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-standardized-errors==0.14.1
et_xmlfile==2.0.0
gunicorn==23.0.0
Markdown==3.7
mypy-extensions==1.0.0
numpy==2.3.2
openpyxl==3.1.5
packaging==24.2
pandas==2.3.1
pandas-stubs==2.3.0.250703
pathspec==0.12.1
phonenumbers==9.0.10
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
